+++ wand/api.py
@@ -89,6 +89,9 @@ library.MagickGetImageFormat.restype = ctypes.c_char_p
 
 library.MagickSetImageFormat.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
 
+library.MagickToMime.argtypes = [ctypes.c_char_p]
+library.MagickToMime.restype = ctypes.POINTER(ctypes.c_char)
+
 library.MagickGetImageBlob.argtypes = [ctypes.c_void_p,
                                        ctypes.POINTER(ctypes.c_size_t)]
 library.MagickGetImageBlob.restype = ctypes.POINTER(ctypes.c_ubyte)
@@ -181,6 +184,8 @@ elif platform.system() == 'Darwin':
     libc = ctypes.cdll.LoadLibrary('libc.dylib')
 else:
     libc = ctypes.cdll.LoadLibrary('libc.so.6')
+
+libc.free.argtypes = [ctypes.c_void_p]
 
 libc.fdopen.argtypes = [ctypes.c_int, ctypes.c_char_p]
 libc.fdopen.restype = ctypes.c_void_p